+++ h2gis-functions/src/main/java/org/h2gis/functions/factory/H2GISFunctions.java
@@ -350,7 +350,8 @@ public class H2GISFunctions {
                 new ST_AsOverpassBbox(),
                 new ST_M(),
                 new ST_TriangulatePolygon(),
-                new ST_MaximumInscribedCircleRadius()
+                new ST_MaximumInscribedCircleRadius(),
+                new ST_SimplifyPolygonHull()
         };
     }