+++ docs/CHANGELOG.md
@@ -1,1 +1,3 @@
 ## Changelog for v2.2.4
+
+- exclude graalvm native image during the build