+++ docs/CHANGELOG.md
@@ -1,1 +1,3 @@
-## Changelog for v2.2.4
+## Changelog for v2.2.5
+
+- Exclude graalvm compiler in the builds